Use feature attributes on map download layer
The map download layer used a part of the vector feature id as
reference and stored additional information in a separate
array. The id format has changed in the new OpenLayers version.

The vector feature supports own attributes. This patch uses
the attributes instead of the id to store additional information.